Distribute jobs with multiprocessing (#19)
* Add badges to README.md
* Distribute jobs with multiprocessing
