+++ hough/__init__.py
@@ -1,1 +1,3 @@
 __version__ = "0.2.0"
+
+WINDOW_SIZE = 150
